createWaterfallLoop implementation needs to be shared  - between LowerSubgroupOperations and BuilderImpl(s)
